Replace svn:externals with submodule, symlink, etc
We only need a portion of css3-foundation, so rather than put the
entire repo into the code, place the submodule at the root level
and then symlink just the portion we need.

Since no longer plan on maintaining js libs as part of the
css3-foundation, check these in as versioned files, rather than
using a submodule.